use floats just in case
Bench 8018277

+++ src/history.hpp
@@ -12,7 +12,7 @@ using ContHistEntry = std::array<std::array<std::array<i32, 64>, 6>, 2>;
 using ContHistory   = std::array<std::array<std::array<ContHistEntry, 64>, 6>, 2>;
 // king can't get captured
 using CaptHistory       = std::array<std::array<std::array<std::array<i32, 64>, 6>, 6>, 2>;
-using CorrectionHistoryEntry = std::array<i64, 4>;
+using CorrectionHistoryEntry = std::array<i32, 4>;
 using CorrectionHistory = std::array<std::array<CorrectionHistoryEntry, 16384>, 2>;
 
 constexpr i32 HISTORY_MAX                     = 16384;
